Make the installed slab allocator arena-aware for realloc
otlp_install_slab_allocator passed realloc straight through to the
previous allocator, so any arena-served allocation that later grew
via otlp_realloc - e.g. the HTTP response buffer (otlp_malloc(4096)
then doubling growth) with slot_size >= 4096 - was handed to libc
realloc on a pointer libc never allocated: undefined behavior
that aborts under macOS libmalloc and corrupts the glibc heap.

The wrapped allocator now MOVES arena pointers: allocate at the
new size, copy min(slot_size, n) bytes, return the slot. Any slot
size is now safe; sizing guidance (and the quickstart/cookbook
examples, now 256/512) is purely about hit rate. Pinned by
prop_slab_global_realloc_growth, which aborts the process on the
pre-fix code.

Also fixes two sign-conversion warnings in test_http_parser.c.

+++ tests/test_http_parser.c
@@ -457,7 +457,7 @@ test_send_stall_times_out(void)
 
 		clock_gettime(CLOCK_MONOTONIC, &t0);
 		deadline_ms = (uint64_t) t0.tv_sec * 1000 +
-			t0.tv_nsec / 1000000 + 2000;
+			(uint64_t) (t0.tv_nsec / 1000000) + 2000;
 		for (i = 0; i < 5000000; i++)
 		{
 			struct timespec now;
@@ -470,7 +470,7 @@ test_send_stall_times_out(void)
 				break;
 			clock_gettime(CLOCK_MONOTONIC, &now);
 			if ((uint64_t) now.tv_sec * 1000 +
-					now.tv_nsec / 1000000 >
+					(uint64_t) (now.tv_nsec / 1000000) >
 				deadline_ms)
 				break;
 			if (st == OTLP_ERR_WOULDBLOCK)
